+++ vite.config.js
@@ -2,12 +2,17 @@ import { defineConfig } from 'vite';
 import wasm from "vite-plugin-wasm";
 import topLevelAwait from "vite-plugin-top-level-await";
 import tailwindcss from '@tailwindcss/vite';
+import react from '@vitejs/plugin-react';
 
 
 export default defineConfig({
   plugins: [
+    react(),
     wasm(),
     topLevelAwait(),
     tailwindcss()
-  ]
+  ],
+  optimizeDeps: {
+    include: ['@dexhunterio/swaps', 'react', 'react-dom']
+  }
 });